Started inventory screen
Made inventory UI game object, which contains the script DrawInventory.
DrawInventory will update the GUI with the inventory screen whenever
the key ‘I’ is pressed. To close the screen press ‘Q’ or ‘esc’. The
inventory screen shows the icons of current items in the inventory and
empty slot icons so far.